+++ CHANGELOG.md
@@ -7,7 +7,7 @@
   - Like with temperature `AffineUnit`s, call `dB(3)` or (`3dB` on the command-line) to use the scale; multiplication works as usual
   - Decibel-units are also supported, e.g. `horsepower @ dB(kW)`
 - Time-based measures and units now interoperable with `datetime.timedelta`
-- Added 54 new units:
+- Added 57 new units:
   - **Cooking**: `tablespoon`, `teaspoon`, `dessertspoon`
   - **Astronomy**: `solar_mass`, `solar_radius`, `solar_irradiance`, `solar_luminosity`, `solar_temperature`, `solar_mass_parameter`, `earth_radius_equatorial`, `earth_radius_polar`, `jupiter_radius_equatorial`, `jupiter_radius_polar`, `earth_mass_parameter`, `jupiter_mass_parameter`
   - **Scientific**: `amagat`, `dalton`, `torr`, `bubnoff`, `eotvos`
@@ -16,7 +16,7 @@
     - radioactive: `curie`, `rutherford`, `thomson`, `roentgen`, `Mache`
     - misc: `spat_length`, `foot_lambert`,
   - **Conventional**:
-    - logarithmic: `decibel`
+    - logarithmic: `bel`, `decibel`, `neper`
     - volume: `gill`, `quart`
     - mass: `scruple`, `apdram`, `apounce`, `appound`, `pennyweight`
     - mechanical: `mach`, `lbf`, `slug`, `psi`, `foot_pound`, `pound_foot`, `kip`